+++ src/main/java/com/example/flightplanning/repository/AirportRepository.java
@@ -3,5 +3,5 @@ package com.example.flightplanning.repository;
 import com.example.flightplanning.entity.Airport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AirportRepository extends JpaRepository<Airport, Long> {
+public interface AirportRepository extends JpaRepository<Airport, Integer> {
 }
